tests/lib/ansible_helper: Move ansible code to extra file
Moving ansible related code into its own file (and out of the hardware
releated code) makes the code more structured.
